+++ .gitignore
@@ -1,2 +1,3 @@
 # Ignores root directory 
 root
+.vscode
