+++ assets/js/script.js
@@ -8,3 +8,7 @@ fetch("./modules/read_directories.php?file")
   .catch(function (err) {
     console.log("Fetch problem: " + err.message);
   });
+
+// $("i[data-file]").click(function () {
+//   // console.log("Holaaa");
+// });
