+++ modules/search.php
@@ -1,0 +1,10 @@
+<div class="fm-search">
+    <div class="mb-0">
+        <div class="input-group input-group-lg">
+            <span class="input-group-text bg-transparent">
+                <i class="fa fa-search"></i>
+            </span>
+            <input type="text" class="form-control" placeholder="Search the files">
+        </div>
+    </div>
+</div>
